Cleanup code and improve latency / frame pacing
