+++ desktop/src/main.rs
@@ -1,5 +1,6 @@
 use std::process::exit;
-use std::time::{Duration, Instant};
+use std::time::Instant;
+
 use tracing_subscriber::EnvFilter;
 use winit::event_loop::EventLoop;
 
@@ -55,23 +56,6 @@ fn main() {
 	};
 
 	tracing::info!("Cef initialized successfully");
-
-	let rendering_loop_proxy = event_loop.create_proxy();
-	let target_fps = 120;
-	std::thread::spawn(move || {
-		loop {
-			let last_render = Instant::now();
-
-			let result = futures::executor::block_on(DesktopWrapper::execute_node_graph());
-			let _ = rendering_loop_proxy.send_event(CustomEvent::NodeGraphExecutionResult(result));
-
-			let frame_time = Duration::from_secs_f32((target_fps as f32).recip());
-			let sleep = last_render + frame_time - Instant::now();
-
-			let _ = rendering_loop_proxy.send_event(CustomEvent::ScheduleBrowserWork(Instant::now()));
-			std::thread::sleep(sleep);
-		}
-	});
 
 	let mut winit_app = WinitApp::new(cef_context, window_size_sender, wgpu_context, event_loop.create_proxy());
